requirements.txt
bitarray~=1.8.1
click==7.1.2
cycler==0.10.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
murmurhash==1.0.5
numpy==1.20.1
Pillow==8.1.2
pybloom-live==3.0.0
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
redis==3.5.3
six==1.15.0
Werkzeug==1.0.1

WTForms~=2.3.3